move SetResponseHandler and RecordHttpHandler into new namespace to match their folder
