+++ test/Datadog.Trace.OpenTracing.IntegrationTests/OpenTracingSendTracesToAgent.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using System.Net;
 using Datadog.Trace.Agent;
 using Datadog.Trace.TestHelpers;
+using Datadog.Trace.TestHelpers.HttpMessageHandlers;
 using Xunit;
 
 namespace Datadog.Trace.OpenTracing.IntegrationTests
